Add support for multiple button devices
This adds support for event 0x3A, button, in the form of
a regular sensor. The button used has four buttons, and thus
this also adds support for multiple measurements, as described
in the BTHome v2 format documentation ( https://bthome.io/format/ ):

"If you want to send multiple measurements of the same type,
e.g. three temperatures, you can just add multiple measurements
of the same type to the payload. A postfix will be added to the
measurement name (e.g. temperature_2) in the order of which you
define the measurements."

As a side effect, this now also happens to any supported multiple
measurements sensors. The BTHome documentation is not consistent
with how multiple measurements are handled in bthome-ble
( https://github.com/Bluetooth-Devices/bthome-ble/ ).

The button used sends 0xFE when held down, which is not consistent
with the BTHome v2 specification. It is silently converted to 0x80,
which is the proper code for held buttons.